Feat/#25/홈화면 병합
